Separate local from global bin installs
Also, have shellfuncs only display an error if something
went wrong with the install. This is because I found it annoying
to see the shellfuncs success message and the dotfiles
success message.

Have each init script output some success message. This is so
I am reminded of what else is going on when dotfiles are installed.