Refactor dotfiles
The install script has now been replaced with a makefile. This
was done in terms of rescoping of the project and
refactorment. The makefile just installs the dotfiles, and that is
it.

The makefile uses stow as a means of dotfile management,
meaning technically there is somewhat less code to manage now.

Codium settings will now be synced to GitHub instead of
managing them via git. Due to these settings changing a lot,
I figured it would probably be best not in a git repo.

Other dotfiles have been refactored, .bashrc got a good
amount refactored due to actually going through what all
the contents did. .gitconfig now includes a local gitconfig as
well incase I wish to add public key for signing commits.