Integrate notion of shell templates into dotfiles
Main reason I wanted this is because of some dotfiles
sharing the name of a local version of that dotfile that would
not be tracked by version control.

While I could have left it hard coded, I decided to have make
control that information and evaluate dotfiles as needed to
have that information.