Make application configs more consistent
When trying to add a new application (and by extension configuration)
I noticed that the repo was somewhat inconsistent in how this was done.

Mainly, .ssh really should have been ssh and the addition of new
applications shouldn't call for a duplication of pathing to
install/uninstall.

Instead, from this point I will assume any stowfile/dotfile will share
the common root directory of $HOME and from there pathing to the
stowfile/dotfile will be based on the directory structure in the
respective app dir.

The main find expression has been refactored. I do not recall why
generated dotfiles needed it own foreach loop. That and the file
recorded is of the form <app>/<path to app config>.
The main comment has been expanded on and some other smaller things
shuffled.