Add jenkins icon for .jenkins directory for now
In the event that vscode-icons supports this natively in the future,
then I will drop my copy(s) of the file icon from this repo.

I should note that the icon pathing may change if I end up changing the
vsicons.customIconFolderPath in my VSCodium settings.

Stow needed to be append with --no-folding so only leafs are symlinked.
As I expect this folder to exist upon codium being installed, but others
that are non-existant will still be created before symlinking.