Set a more consistent PATH across dev machines
Related to 91dd1567 with regard to /sbin. Distributions can manipulate
the PATH to include non-standard directories (e.g. /usr/games,
/snap/bin) and exclude standard directories (e.g. Debian does this for
/sbin and its variants for non-root users). The desire is for pathing to
be consistent across my development machines.

To address the latter, I decided to set the PATH to some common path.
The value used came from what the bash man page considered to be a
common value. For the former, I wish to retain the distribution's
non-standard directories set in the PATH as a distribution could utilize
the non-standard directories for binaries (e.g. cowsay in /usr/games)
that come from a distribution's package manager.