Continue work of 91dd1567 for the .bashrc file
The first case statement checking for an interactive shell has been
replaced with a simple if expression. Currently, if all that is desired
is that the interactive option is set on a shell, then said logic can be
represented more simply with an if expression.

Aliases, shell functions, and shell options being set (via shopt) have
now been moved relative to where I would put these in a formal shell
script. At least, their new locations have been inspired by such.

Instead of using the -z and -n test expressions for checking the
existence (or lack thereof) of kubectl and minikube. I feel using the -x
expression is closer to the intention that I look for when looking for
executables. In that they not only exist in the PATH, but are files I
can invoke at the shell.

To limit the number of subshells spawned when loading this startup file,
if it is with great certainty that a executable will come from the
distribution, then paths can be hard coded with /usr/bin (e.g. lesspipe,
dircolors).

lesspipe no longer sets the SHELL statically before being invoked. This
follows how the Debian man page says to invoke the executable. That, and
the SHELL environment variable should already be set according to the
shell you are using.

The aliases ls, grep, fgrep, and egrep original locations I disagree
with because the existence (or lack thereof) of dircolors doesn't affect
if the aliases can output colored output to the terminal. That should be
determined by examining the kind of terminal used to load this startup
file. This logic is already being performed and discussed in the
paragraph below.

The logic for setting a fancy prompt has been reduced down to a simple
if/else expression. This was done because I felt the case statement was
unnecessary given there would be only one pattern that could be more
simply written as an if/else expression. tput is used to query the
terminfo database to determine color support. This was done because the
logic more is accurate in determining if a terminal has color support.
This is true for Tektronix-like terminals.

File paths not quoted have now been corrected to include quotes. This is
to match my personal conventions.