Move pattern rule closer to prereqs' declaration
I believe the idea at 0022a001 was that I would have more
pattern/implicit rules to add eventually in the Makefile. Since I do not
currently think that will be the case anytime soon, I believe it would
be more appropriate to define the pattern rule closer to where the
prerequisites are declared. The positioning was inspired by a Trello
card that pointed to
https://github.com/cavcrosby/debcomprt/commit/f6cae6d.