Add shell tracing to rmplain-files target
I found that enabling shell tracing and setting the PS4 environment
variable to be empty would get most (if not all) of the way there for
mimicking what the echo statements were doing for me in the for loops.
This reduces the amount of duplicate code needed and shows a more
accurate runtime tracing (e.g. rm --force will only show if a ${HOME}
dotfile is a symlink to a stowfile).

rmplain-files has been removed as a prerequisite from the install target
because since that target/prerequisite always runs its recipe, its
output bloats the install target's output. Which isn't desired now with
these new changes.