Add the tmux terminal multiplexer
VIRTUAL_ENV is unset in the global environment because when starting a
new pane, if the Python version is set the same according to by pyenv
then pyenv will assume the Python virtual environment to already be
activated. This results in the prompt shell (PS1) not containing the
current Python virtual environment.