Refactor the Makefile
The substitution reference expression has been replaced with a patsubst
function call because the simpler notation of a substitution reference
isn't that much more beneficial to me and I'll have an easier time
remembering what this expression is called (a patsubst function call).

Further unnecessary comments have been removed (or changed) because they
did not provide any additional value as the related code is
self-explanatory.

stowfiles is now pkg_file_paths as what I meant by a "stowfile" was a
file within a package directory, so package (pkg) file is a more
appropriate term.