Add the Docker CLI config.json
Because of the shell command substitution performed for defining the
ENCODED_DOCKER_HUB_AUTH_STR variable, the common configuration file
heredoc will need to shut off doing variable expansion.

With the (decoded) ENCODED_DOCKER_HUB_AUTH_STR variable format being
<username>:<api_token>, I decided that since my username is a publicly
known value and for the convenience factor, that for now I will hardcode
that value. An API token must be used as part of the auth string because
Docker Hub will not allow password authentication via Docker CLI when
the Docker Hub account has 2FA enabled.

Since the config.json has been added as a shell template, it is no
longer just dotfiles that are being created from shell templates. Hence
generically, we can say (stow) package files can be created from shell
templates. Instead of searching (and filtering) for package files that
are shell templates, putting the results into variables, I felt it would
be simpler to have all package files be a prerequisite and use a
terminal match-anything rule to run for those that have a shell
template.

The logic in the rmplain-files rule had a bug for when if a file did not
exist, then the rm command would still run. I've corrected the logic to
first ensure the file does exist before checking if it is a symbolic
link.

The docker alias has been added because 'docker login' has an issue
replacing the actual config.json file with an updated temporary
"config.json" file due to pathing of where the Docker CLI command is
invoked. To circumvent the issue, the Docker CLI command must be invoked
within the ${HOME}/.docker directory.