Create Makefile target to set pkg-files file modes
It was brought to my attention that permissions set on the stow package
files that are created from shell templates can be overly permissive
(e.g. world read on .netrc). Hence, I decided to create a Makefile
target to set the desired file modes permissions for package files that
are created from shell templates or for packages files that have a
recommended file mode to use (e.g. authorized_keys).

The ssh authorized_keys and config package files permissions derived
from what was recommended in ssh(1) for 'OpenSSH_8.9p1'. config was not
made group-writable because I don't have a reason to enable such
permission.