Deduplicate pyenv & rbenv entries in tmux's PATH
Changes were made to remove the duplicate pyenv-virtualenv & rbenv
entries in a tmux global environment's PATH. These duplicate entries
were not found in the shell that invokes tmux.

A duplicate pyenv-virtualenv shim path is checked for but not a pyenv
shim path because as of pyenv 2.3.31, 'pyenv init -' expands with logic
that does remove any duplicate pyenv shim path entries.

The default-command tmux command is set to use the value from the SHELL
environment variable because while that environment variable is used by
default, tmux invokes the shell as a login shell. This isn't desired for
tmux sessions ran on my development machines because that means
.bash_profile/.profile is read in again by the shell. Explicitly setting the
command this way invokes a non-login shell.