Remove the docker alias from .bashrc
This was done because, although the issue noted in cd0bd79 is still
present, the alias causes issues when building images. Changing the
working directory before any 'docker build' command means specifying a
build context of "${PWD}" would actually mean the directory the alias
changes to, and not the one my shell is currently in.

The command from the alias has been added to the chktooling function
because, as previously mentioned, the issue with the 'docker login'
command is still present.