Update the global git/.gitignore file
This was done to move away from copying a template verbatim for this
.gitignore file and to make it my own.

The commented .vscode/ patterns that are negated, and .vsix pattern were
removed because they are unnecessary. The .history/ pattern was removed
because, upon investigation, this directory would be created when using
the Local History Visual Studio Code plugin. A plugin that I currently
don't use.