Hardcode the GPG signing key ID within .gitconfig
This was done because I don't currently use a different signing key and
that the signing key ID is a publicly known value when looking at a
commit's verified status on GitHub.

.gitconfig is no longer a shell template because there are no shell
variables that need to be evaluated before using it. Also, because since
this file is no longer a shell template, its file mode does not need to
be set as part of the chmod-files target.