Add genpass shell function to .bashrc
sed is used instead of printf to append a final newline to stdout
because when using sed, the output looks fine even if a shell function
argument isn't passed in for the passlen parameter. Whereas with printf,
presuming a simple '%s\n' format string, the same scenario causes two
newlines to be appended in the final output.