Add slashes to relative paths in .gitignore
This was done because after reading gitignore(5), I realized I had some
misconceptions about the behaviors of some search patterns. In that,
patterns without a slash or directory separator at the beginning or
middle are applied elsewhere below the .gitignore file. For example,
the '.conf' pattern would match .conf, foo/.conf, and foo/bar/.conf
files. This isn't the intended behavior I desire, and instead I intended
on matching just .conf, hence the addition of a directory separator.

Technically, only the '.conf' pattern needed the addition of a beginning
directory separator, but for consistency sakes, this was applied to
patterns that already had a middle directory separator as well.