Add ssh_keygen shell function to the .bashrc file
This function was created to provide me with a reference on creating new
SSH key pairs for my development machines and to set a format for the
keys' comment. ED25519 key pairs will be generated instead of RSA key
pairs as this is the recommended SSH key type on the interweb, which
provides the same level of security as RSA key pairs and shorter keys.