Add the git/.git-credentials file
Despite the STORAGE FORMAT section of git-credential-store(1)
recommending that this file not be viewed or edited with an editor,
I still want to manage this file under this repository. To add, there
isn't a reason provided for this recommendation. Aside that these
credentials could be exposed via shoulder surfing when viewed or edited
with an editor, I don't currently think there is a compelling
reason not to manage this file under this repository.

The .git-credentials file does not contain an ending newline because,
according to the same section, this kind of lines is not allowed.