Migrate much of .profile.shtpl into a .rc file
This was done because I felt that the commands from the .profile.shtpl
file could be run from a .bashrc-like file, except, that said file would
be considered generic for any Bourne shell. Hence, I've created the .rc
file to fill that purpose. This also avoids the modifying PATH
environment variable during the creation of a graphical desktop session,
which I've noticed can lead to issues. One of which is described here,
https://github.com/cavcrosby/desktop-configs/commit/574f3a3. The .rc
file description was inspired by the .profile.shtpl file.

There is an additional benefit that this will enable things to work with
the GNOME Display Manager when it runs Wayland sessions.