Add the pre-commit framework to the repository
Despite being resistant to adding any language (e.g. Python, Ruby)
package dependencies to the repository, I am fine with relaxing this
stance for pre-commit because it is vital for maintaining most of my
repositories. I did explore using the pre-commit package from my
distribution's repositories; however, the version (2.17.0) it has runs
with an error. A remedy suggested for the error is to just upgrade the
pre-commit version (as mentioned here,
https://github.com/psf/black/issues/4065#issuecomment-1821277094).

The shellcheck directives have been changed to pass the shellcheck
pre-commit hook. Previously, these directives were written to pass the
shellcheck VSCode extension, but for now, it is intended that I will
solely rely on the former. This is because there isn't a native way to
set the version of shellcheck that should be used to lint the shell
code, not counting the version used by its pre-commit hook.